+++ drizzle/0019_normal_the_leader.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `sources` ADD `probe_detect_black_screen` integer DEFAULT true NOT NULL;